redesigned layout and made login/signup pages fixed
