+++ App/frontend/src/App.css
@@ -4,23 +4,25 @@
   background-size: cover;
 }
 
-.form-container {
-  /*background-color: #eeebdd; */
+/*.form-container {
   display: flex;
   justify-content: center;
   align-items: center;
   width: 100vw;
-  height: 80vh;
-}
+  height: 70vh;
+}*/
 
 .form {
+  position: fixed;
+  z-index: 2;
+  top: 20%;
+  left: calc(50% - 210px);      /*%50 - width/2 */
   width: 420px;
   box-shadow: rgb(0 0 0 / 16%) 1px 1px 10px;
   padding-top: 30px;
   padding-bottom: 20px;
   border-radius: 8px;
   background-color: white;
-  z-index: 2;
 }
 
 .form-content {
@@ -30,7 +32,7 @@
 
 .form-title {
   text-align: center;
-  font-family: 'Inter', sans-serif;
+  font-family: "Inter", sans-serif;
   margin-bottom: 1em;
   font-size: 24px;
   color: rgb(34, 34, 34);
@@ -45,7 +47,7 @@
 
 .wrapper {
   /*background-color: #eeebdd;*/
-  font-family: 'Inter', sans-serif;
+  font-family: "Inter", sans-serif;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -65,6 +67,10 @@
   background-color: #e3f0fa;
 }
 
+.btn-outline-primary:active {
+  background-color: #85bbdd !important;
+}
+
 .googleButton {
   margin-left: 3em;
 }
@@ -76,12 +82,3 @@
 .bg-secondary {
   background-color: yellow !important;
 }
-
-/*.rightImage {
-  background-image: url("img_tree.png");
-  background-repeat: no-repeat;
-  background-position: right top;
-  margin-right: 200px;
-  background-attachment: scroll;
-}*/
-
