+++ App/frontend/src/App.js
@@ -3,13 +3,15 @@ import "./App.css";
 import React from "react";
 import { Routes, Route } from "react-router-dom";
 import Homepage from "./pages/Homepage";
-// import Layout from "./layout/Layout";
+import Layout from "./layout/Layout";
 
 function App() {
   return (
-    <Routes>
-      <Route path="/" element={<Homepage />} />
-    </Routes>
+    <Layout>
+      <Routes>
+        <Route path="/" element={<Homepage />} />
+      </Routes>
+    </Layout>
   );
 }
