+++ App/frontend/src/layout/Layout.module.css
@@ -1,5 +1,5 @@
 .main {
-  margin: 3rem auto;
+  margin: 2rem auto;
   width: 90%;
-  max-width: 85rem;
+  max-width: 82rem;
 }
