+++ App/frontend/src/pages/Homepage.module.css
@@ -1,0 +1,4 @@
+.deneme {
+    font-size: 36px;
+    color: white;
+}
